requirements.txt
certifi==2018.4.16
chardet==3.0.4
defusedxml==0.5.0
Django==2.0.6
djangorestframework==3.8.2
idna==2.6
oauthlib==2.1.0
PyJWT==1.6.4
python3-openid==3.1.0
pytz==2018.4
requests==2.18.4
requests-oauthlib==1.0.0
six==1.11.0
urllib3==1.22
